Port to ejabberd 13.10
The most user visible change is that the configuration is now done using
YAML syntax instead of Erlang syntax. These brings some restrictions
causing the transition requiring translating the configuration file
manually.

An example of a old vs new configuration:

Old:

{listen, [
          {{8088, {127, 0, 0, 1}}, ejabberd_http,
           [{request_handlers, [{["api"], mod_restful}]}]}
         ]}.

{modules, [
       {mod_restful, [
        {api,
         [
          {["admin"], mod_restful_admin, [
           {key, "secret"},
           {allowed_commands, [register, unregister]}
          ]},
          {["register"], mod_restful_register, [{key, "secret"}]}
         ]}
       ]}
      ]}.

New:

listen:
  -
    port: 8088
    ip: "::"
    module: ejabberd_http
    request_handlers:
      "/api": mod_restful

modules:
  mod_restful:
    api:
      - path: ["admin"]
          module: mod_restful_admin
        params:
          key: "secret"
          allowed_commands: [register, unregister]
      - path: ["register"]
          module: mod_restful_register
        params:
          key: "secret"

Most of the porting involved changing strings to binary.

All tests in tests/ and priv/tests pass but this commit is not tested
further than that.